version update

+++ createContainer.sh
@@ -8,7 +8,12 @@ DATA_DIR=data
 
 [ -d $DATA_DIR ] || mkdir $DATA_DIR
 
-docker stop $containerName
-docker rm $containerName
-
-docker run -d --hostname $vhostName --name $containerName -p 8080:15672 -p 5672:5672 -v `pwd`/$DATA_DIR:/var/lib/rabbitmq/mnesia/rabbit@$vhostName $imageName
+if pullLatestDocker.sh -i $imageName
+then
+    docker stop $containerName
+    docker rm $containerName
+    
+    docker run -d --hostname $vhostName --name $containerName -p 8080:15672 -p 5672:5672 -v `pwd`/$DATA_DIR:/var/lib/rabbitmq/mnesia/rabbit@$vhostName $imageName
+else
+    echo "image already latest, so no update"
+fi
